requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0
django-countries==7.5.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
